+++ lib/wine_bouncer/oauth2.rb
@@ -62,12 +62,10 @@ module WineBouncer
       unless valid_doorkeeper_token?(*scopes)
         if !doorkeeper_token || !doorkeeper_token.accessible?
           error = Doorkeeper::OAuth::InvalidTokenResponse.from_access_token(doorkeeper_token)
-          # TODO: localization and better error reporting
-          raise WineBouncer::Errors::OAuthUnauthorizedError, 'unauthorized'
+          raise WineBouncer::Errors::OAuthUnauthorizedError, error
         else
           error = Doorkeeper::OAuth::ForbiddenTokenResponse.from_scopes(scopes)
-          # TODO: localization and better error reporting
-          raise WineBouncer::Errors::OAuthForbiddenError, "missing permissions"
+          raise WineBouncer::Errors::OAuthForbiddenError, error
         end
       end
     end